Move Mask class and _translate_mask_method to aperture core module
